requirements.txt
numpy
opencv-contrib-python
PyYAML
ultralytics
flask
flask-cors
pyserial
